% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-description_analysis.R
\name{Barchart.Pango}
\alias{Barchart.Pango}
\title{generate barchart}
\usage{
Barchart.Pango(dataset, aim_col, col_num)
}
\arguments{
\item{dataset}{the dataset}

\item{aim_col}{the aim col}

\item{col_num}{the number of cols}
}
\description{
generate barchart
}
