% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-change_block_pango_S3.R
\name{Change_Block_Pango}
\alias{Change_Block_Pango}
\title{Change Block Pango version S3 class}
\usage{
Change_Block_Pango(
  stock_symbol,
  start_date,
  end_date,
  block_period,
  block_dataset,
  inflection_point,
  block_volatility,
  block_order,
  block_direction,
  block_location,
  change_state
)
}
\arguments{
\item{stock_symbol}{the stock symbol}

\item{start_date}{the start date}

\item{end_date}{the end date}

\item{block_period}{the block period}

\item{block_dataset}{the block dataset}

\item{inflection_point}{the inflection point}

\item{block_volatility}{the block volatility}

\item{block_order}{the block order}

\item{block_direction}{the block direction}

\item{block_location}{block_location}

\item{change_state}{the change state}
}
\description{
Change Block Pango version S3 class
}
