% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-additional_dataset.R
\name{.add_moving_average_information}
\alias{.add_moving_average_information}
\title{add moving average information}
\usage{
.add_moving_average_information(origi_dataset, ma_col)
}
\arguments{
\item{origi_dataset}{the original dataset}

\item{ma_col}{the moving average column}
}
\description{
add moving average information
}
