% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-additional_dataset.R
\name{.annotate_last_inflection_point_and_interval}
\alias{.annotate_last_inflection_point_and_interval}
\title{annotate last inflection point and interval}
\usage{
.annotate_last_inflection_point_and_interval(stock_block, ma_col)
}
\arguments{
\item{stock_block}{the stock block}

\item{ma_col}{the moving average column}
}
\description{
annotate last inflection point and interval
}
