% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-additional_dataset.R
\name{.annotate_moving_average_direction}
\alias{.annotate_moving_average_direction}
\title{annotate moving average direction}
\usage{
.annotate_moving_average_direction(origi_dataset, ma_col)
}
\arguments{
\item{origi_dataset}{the original dataset}

\item{ma_col}{the moving average column}
}
\description{
annotate moving average direction
}
