% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-additional_dataset.R
\name{.generate_additional_dataset}
\alias{.generate_additional_dataset}
\title{generate additional dataset}
\usage{
.generate_additional_dataset(origi_dataset)
}
\arguments{
\item{origi_dataset}{the original dataset}
}
\description{
generate additional dataset
}
