% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-stock_block.R
\name{.generate_block_information_matrix}
\alias{.generate_block_information_matrix}
\title{generate block information matrix}
\usage{
.generate_block_information_matrix(block_ls)
}
\arguments{
\item{block_ls}{the stock block list}
}
\description{
generate block information matrix
}
