% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-stock_block.R
\name{.generate_block_location_sequence}
\alias{.generate_block_location_sequence}
\title{generate block location sequence}
\usage{
.generate_block_location_sequence(origi_dataset, ma_symbol)
}
\arguments{
\item{origi_dataset}{the original dataset}

\item{ma_symbol}{the moving average symbol}
}
\description{
generate block location sequence
}
