% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-stock_block.R
\name{.generate_block_sequence}
\alias{.generate_block_sequence}
\title{generate block sequence}
\usage{
.generate_block_sequence(seq_breaks)
}
\arguments{
\item{seq_breaks}{the sequence breaks}
}
\description{
generate block sequence
}
