% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-stock_block.R
\name{.generate_graphic_matrix}
\alias{.generate_graphic_matrix}
\title{generate graphic matrix}
\usage{
.generate_graphic_matrix(change_blocks, information_matrix)
}
\arguments{
\item{change_blocks}{the change blocks}

\item{information_matrix}{the information matrix}
}
\description{
generate graphic matrix
}
