% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-description_analysis.R
\name{.generate_plotting_cols}
\alias{.generate_plotting_cols}
\title{generate plotting cols}
\usage{
.generate_plotting_cols(original_dataset, ma_symbol)
}
\arguments{
\item{original_dataset}{the original dataset}

\item{ma_symbol}{the moving average symbol}
}
\description{
generate plotting cols
}
