% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-stock_block.R
\name{.generate_stock_block}
\alias{.generate_stock_block}
\title{generate stock block}
\usage{
.generate_stock_block(origi_dataset, ma_symbol)
}
\arguments{
\item{origi_dataset}{the original dataset}

\item{ma_symbol}{the moving average symbol}
}
\description{
generate stock block
}
