% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-stock_block.R
\name{.generate_trend_matrix}
\alias{.generate_trend_matrix}
\title{generate trend matrix}
\usage{
.generate_trend_matrix(graphic_matrix, ma_col = "sma_d5")
}
\arguments{
\item{graphic_matrix}{the graphic matrix}
}
\description{
generate trend matrix
}
