% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-stock_block.R
\name{.generate_trend_state}
\alias{.generate_trend_state}
\title{generate trend state}
\usage{
.generate_trend_state(block_info_mat)
}
\arguments{
\item{block_info_mat}{the stock block}
}
\description{
generate trend state
}
