% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-additional_dataset.R
\name{.verify_year_group}
\alias{.verify_year_group}
\title{verify year group}
\usage{
.verify_year_group(years)
}
\arguments{
\item{years}{the year group to verify}
}
\description{
verify year group
}
