% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generic-stock_database_pango.R
\name{download_stock_dataset_pango}
\alias{download_stock_dataset_pango}
\title{download stock dataset pango}
\usage{
download_stock_dataset_pango(x, ...)
}
\description{
download stock dataset pango
}
