% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generic-stock_database_pango.R
\name{download_stock_dataset_pango.Stock_Database_Pango}
\alias{download_stock_dataset_pango.Stock_Database_Pango}
\title{download stock dataset pango}
\usage{
\method{download_stock_dataset_pango}{Stock_Database_Pango}(
  x,
  stock_symbol,
  year,
  period = "daily",
  adjust = "qfq",
  market = "A",
  ...
)
}
\arguments{
\item{stock_symbol}{the stock symbol}

\item{year}{the year to download}

\item{period}{the data period}

\item{adjust}{the data adjust}

\item{market}{the market symbol}
}
\description{
download stock dataset pango
}
