% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generic-stock_database_pango.R
\name{extract_stock_dataset_pango}
\alias{extract_stock_dataset_pango}
\title{extract stock dataset pango}
\usage{
extract_stock_dataset_pango(x, ...)
}
\description{
extract stock dataset pango
}
