% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generic-stock_database_pango.R
\name{extract_stock_dataset_pango.Stock_Database_Pango}
\alias{extract_stock_dataset_pango.Stock_Database_Pango}
\title{extract stock dataset pango}
\usage{
\method{extract_stock_dataset_pango}{Stock_Database_Pango}(
  x,
  stock_symbol,
  year,
  period = "daily",
  adjust = "qfq",
  market = "A",
  download = FALSE,
  ...
)
}
\arguments{
\item{stock_symbol}{the stock symbol}

\item{year}{the year to download}

\item{period}{the data period}

\item{adjust}{the data adjust}

\item{market}{the market symbol}

\item{download}{whether download from internet}
}
\description{
extract stock dataset pango
}
