% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-stock_block_pango_S4.R
\name{initialize,Stock_Block_Pango-method}
\alias{initialize,Stock_Block_Pango-method}
\title{initialize Stock_Block_Pango}
\usage{
\S4method{initialize}{Stock_Block_Pango}(.Object, original_dataset, ma_symbol)
}
\arguments{
\item{original_dataset}{the original dataset}

\item{ma_symbol}{the moving average symbol}
}
\description{
initialize Stock_Block_Pango
}
