% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-stock_block_pango_S4.R
\docType{class}
\name{Stock_Block_Pango-class}
\alias{Stock_Block_Pango-class}
\title{the class of Stock Block Pango}
\arguments{
\item{change_blocks}{the block list}

\item{information_matrix}{information matrix}

\item{graphic_matrix}{the matrix for plotting}
}
\description{
the class of Stock Block Pango
}
