% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-stock_block_pango_S4.R
\name{Stock_Block_Pango}
\alias{Stock_Block_Pango}
\title{constructor}
\usage{
Stock_Block_Pango(original_dataset, ma_symbol)
}
\arguments{
\item{original_dataset}{the original dataset}

\item{ma_symbol}{the moving average symbol}
}
\description{
constructor
}
