% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-stock_database_pango_S3.R
\name{Stock_Database_Pango}
\alias{Stock_Database_Pango}
\title{Stock Database Pango version S3 class}
\usage{
Stock_Database_Pango(
  db_na = NULL,
  db_path = NULL,
  activate = FALSE,
  message_connect = TRUE,
  message_disconnect = TRUE
)
}
\arguments{
\item{db_path}{the database path}
}
\description{
Stock Database Pango version S3 class
}
