% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-description_analysis.R
\name{Trend_Graph.Pango}
\alias{Trend_Graph.Pango}
\title{trend plot}
\usage{
Trend_Graph.Pango(
  stock_block,
  interval = numeric(),
  interval_width = 100,
  ma_symbol = "sma_d5",
  linewidth = 0.3
)
}
\arguments{
\item{stock_block}{the stock block class}

\item{interval}{the description interval}

\item{interval_width}{the interval width}
}
\description{
trend plot
}
