% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-stock_block_pango_S4.R
\name{Stock_Block_Pango}
\alias{Stock_Block_Pango}
\title{constructor}
\usage{
Stock_Block_Pango(
  stock_database,
  stock_symbol,
  year,
  period = "daily",
  adjust = "qfq",
  market = "A",
  ma_symbol = "sma_d5"
)
}
\arguments{
\item{stock_database}{the stock database}

\item{stock_symbol}{the stock symbol}

\item{year}{the year to download}

\item{period}{the data period}

\item{adjust}{the data adjust}

\item{market}{the market symbol}
}
\description{
constructor
}
