% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-description_analysis.R
\name{Trend_Graph.Pango}
\alias{Trend_Graph.Pango}
\title{trend plot}
\usage{
Trend_Graph.Pango(
  dataset,
  interval = c(1, nrow(dataset)),
  ma_symbol = "sma",
  ma_interval = 5
)
}
\arguments{
\item{dataset}{the dataset}

\item{interval}{the description interval}
}
\description{
trend plot
}
